geotiff tests: consolidate kwarg / signature cluster (#2431)
Fold nine top-level kwarg and signature test files into a new
xrspatial/geotiff/tests/unit/test_signatures.py, organised into six
sections by concern: parameter annotations (#1654, #1705), canonical
reader kwarg order (#1935), experimental / internal-only opt-in gates
(#2352), the photometric kwarg and extra_tags override (#1769), the
gil_friendly deflate kwarg (#1830), and the 2026-05-12 reader / writer
kwarg-behaviour coverage sweep.

GPU gating moves to the shared requires_gpu marker from
_helpers/markers.py, replacing the per-file _gpu_available helpers.
No assertion changed; the consolidated file collects the same 167
tests as the nine source files.

CLUSTER_AUDIT_KWARG.md maps every old file::test to its new
file::test_id and is deleted in a final pre-merge commit per epic #2424.

Tests-only -- no source changes.
